+++ tests/fixtures/controls/dummy_fail_loading_experiment.py
@@ -1,0 +1,4 @@
+from chaoslib.exceptions import InterruptExecution
+
+def before_loading_experiment_control(context: str):
+    raise InterruptExecution("failed to load: {}".format(context))