+++ tests/fixtures/controls/dummy_retitle_experiment_on_loading.py
@@ -1,0 +1,5 @@
+from chaoslib.types import Experiment
+
+
+def after_loading_experiment_control(context: str, state: Experiment):
+    state["title"] = "BOOM I changed it"